pyproject.toml
[tool.poetry]
name = "readme-project"
version = "0.1.0"
description = ""
authors = ["Miklós Koren <koren@localhost>"]
readme = "README.md"
packages = [{include = "readme_project"}]

[tool.poetry.dependencies]
python = "^3.11"
tqdm = "^4.66.1"
pypdf2 = "^3.0.1"
textract = "^1.6.5"
pdf2txt = "^0.7.14"
cerberus = "^1.3.5"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
